creacion de los addons faltantes y sus css

+++ practica2/addon_suma/views.py
@@ -4,8 +4,8 @@ def sumar(request):
     resultado = None
 
     if request.method == "POST":
-        a = int(request.POST.get("a", 0))
-        b = int(request.POST.get("b", 0))
+        a = float(request.POST.get("a", 0))
+        b = float(request.POST.get("b", 0))
         resultado = a + b
 
     return render(request, "addon_suma/sum.html", {"resultado": resultado})
